+++ packages/gatsby-telemetry/src/telemetry.js
@@ -7,7 +7,7 @@ const { join, sep } = require(`path`)
 const isDocker = require(`is-docker`)
 const showAnalyticsNotification = require(`./showAnalyticsNotification`)
 const lodash = require(`lodash`)
-const { getRepositoryId: _getRepositoryId } = require(`./repository-id`)
+import { getRepositoryId as _getRepositoryId } from "./repository-id"
 
 module.exports = class AnalyticsTracker {
   store = new EventStorage()
@@ -196,11 +196,7 @@ module.exports = class AnalyticsTracker {
   }
 
   getDbEngine() {
-    if (process.env.GATSBY_DB_NODES === `loki`) {
-      return `loki`
-    } else {
-      return `redux`
-    }
+    return `redux`
   }
 
   getMachineId() {