+++ www/src/assets/ornaments/index.js
@@ -1,5 +1,11 @@
+import CirclesOrnament from "!raw-loader!./circles-ornament.svg"
 import NewsletterFormOrnament from "!raw-loader!./newsletter-form-ornament.svg"
-import StarOrnament from "!raw-loader!./star-ornament.svg"
 import QuotationMarkOrnament from "!raw-loader!./quotation-mark-ornament.svg"
+import StarOrnament from "!raw-loader!./star-ornament.svg"
 
-export { NewsletterFormOrnament, StarOrnament, QuotationMarkOrnament }
+export {
+  CirclesOrnament,
+  NewsletterFormOrnament,
+  QuotationMarkOrnament,
+  StarOrnament
+}