+++ packages/gatsby-plugin-manifest/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.4.1...gatsby-plugin-manifest@2.4.2) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+## [2.4.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.4.0...gatsby-plugin-manifest@2.4.1) (2020-04-28)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+# [2.4.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.7...gatsby-plugin-manifest@2.4.0) (2020-04-27)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+## [2.3.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.6...gatsby-plugin-manifest@2.3.7) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+## [2.3.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.5...gatsby-plugin-manifest@2.3.6) (2020-04-21)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
+## [2.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.4...gatsby-plugin-manifest@2.3.5) (2020-04-17)
+
+### Bug Fixes
+
+- wrap ignore pattern in quotes ([#23176](https://github.com/gatsbyjs/gatsby/issues/23176)) ([7563db6](https://github.com/gatsbyjs/gatsby/commit/7563db6))
+
+## [2.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.3...gatsby-plugin-manifest@2.3.4) (2020-04-16)
+
+**Note:** Version bump only for package gatsby-plugin-manifest
+
 ## [2.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-manifest@2.3.2...gatsby-plugin-manifest@2.3.3) (2020-03-26)
 
 **Note:** Version bump only for package gatsby-plugin-manifest