+++ packages/gatsby-plugin-sharp/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.1...gatsby-plugin-sharp@2.6.2) (2020-05-07)
+
+### Bug Fixes
+
+- **gatsby-image:** Fix fluid not respecting maxWidth/maxHeight ([#17006](https://github.com/gatsbyjs/gatsby/issues/17006)) ([ad7cd6b](https://github.com/gatsbyjs/gatsby/commit/ad7cd6b))
+
+## [2.6.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.6.0...gatsby-plugin-sharp@2.6.1) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+# [2.6.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.7...gatsby-plugin-sharp@2.6.0) (2020-04-27)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.5.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.6...gatsby-plugin-sharp@2.5.7) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-plugin-sharp
+
+## [2.5.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.5...gatsby-plugin-sharp@2.5.6) (2020-04-17)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** error if we can't determine dimensio… ([#23156](https://github.com/gatsbyjs/gatsby/issues/23156)) ([e11c970](https://github.com/gatsbyjs/gatsby/commit/e11c970))
+- wrap ignore pattern in quotes ([#23176](https://github.com/gatsbyjs/gatsby/issues/23176)) ([7563db6](https://github.com/gatsbyjs/gatsby/commit/7563db6))
+
+## [2.5.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.4...gatsby-plugin-sharp@2.5.5) (2020-04-16)
+
+### Bug Fixes
+
+- **gatsby-plugin-sharp:** trim option failed ([#22997](https://github.com/gatsbyjs/gatsby/issues/22997)) ([0558443](https://github.com/gatsbyjs/gatsby/commit/0558443))
+
+## [2.5.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.3...gatsby-plugin-sharp@2.5.4) (2020-04-04)
+
+### Bug Fixes
+
+- **dependencies:** bump potrace version ([#22803](https://github.com/gatsbyjs/gatsby/issues/22803)) ([7a8f016](https://github.com/gatsbyjs/gatsby/commit/7a8f016))
+
 ## [2.5.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-sharp@2.5.2...gatsby-plugin-sharp@2.5.3) (2020-03-24)
 
 ### Bug Fixes