+++ packages/gatsby-plugin-subfont/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.2.1...gatsby-plugin-subfont@2.2.2) (2020-05-07)
+
+**Note:** Version bump only for package gatsby-plugin-subfont
+
+## [2.2.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.2.0...gatsby-plugin-subfont@2.2.1) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-plugin-subfont
+
+# [2.2.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.1.4...gatsby-plugin-subfont@2.2.0) (2020-04-27)
+
+**Note:** Version bump only for package gatsby-plugin-subfont
+
+## [2.1.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.1.3...gatsby-plugin-subfont@2.1.4) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-plugin-subfont
+
+## [2.1.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.1.2...gatsby-plugin-subfont@2.1.3) (2020-04-17)
+
+### Bug Fixes
+
+- wrap ignore pattern in quotes ([#23176](https://github.com/gatsbyjs/gatsby/issues/23176)) ([7563db6](https://github.com/gatsbyjs/gatsby/commit/7563db6))
+
+## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.1.1...gatsby-plugin-subfont@2.1.2) (2020-04-16)
+
+**Note:** Version bump only for package gatsby-plugin-subfont
+
 ## [2.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-subfont@2.1.0...gatsby-plugin-subfont@2.1.1) (2020-03-23)
 
 **Note:** Version bump only for package gatsby-plugin-subfont